refactor: use useAtomValue for read-only atoms
