+++ src/routes/dashboard/BoxChart/BoxChart.tsx
@@ -1,7 +1,7 @@
 import { useAtom } from 'jotai';
 import type { RectShape } from '../../../models/shape';
 import { BoxButton } from '../../../components';
-import { useSelectController, selectedItemsAtom } from '../atoms';
+import { useSelectController, selectedItemsAtom } from '../concerns';
 
 const BoxChart = ({ className = '', data }: { className?: string; data: RectShape[] | undefined }): JSX.Element => {
   const [selectedItems] = useAtom(selectedItemsAtom);
