+++ src/routes/dashboard/hooks/useActualValues.ts
@@ -1,6 +1,6 @@
 import { useAtom } from 'jotai';
 import { useMemo } from 'react';
-import { selectedItemsAtom } from '../atoms';
+import { selectedItemsAtom } from '../concerns';
 import { useShapes } from '../query/useShapes';
 
 export interface ActualValuesType {
